Moved Syntax source into a separate file
